move string -> color fns into lib.rs
